Separate react components into individual files
